+++ packages/orq-rc/src/orq_ai_sdk/models/updatepromptsnippetop.py
@@ -1908,7 +1908,7 @@ class UpdatePromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-06T09:45:53.820Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-06T14:05:20.313Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")